+++ contextualized_topic_models/networks/inference_network.py
@@ -1,5 +1,3 @@
-"""PyTorch class for feed foward inference network."""
-
 from collections import OrderedDict
 from torch import nn
 import torch